bin/conch| move the "rack" and "racks" commands into their own files
